FlaskNosql/requirements.txt
cachelib==0.1
click==7.1.2
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Session==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
six==1.17.0
Werkzeug==1.0.1